Numerous improvements:
* Better block list - filter IP addresses AND text
* Better error tolerance - don't fail on bad JSON
* IP address anonymizer
* Test scripts with test data
* Console output about submissions and ban activity

The IP address anonymizer works by hashing IP addresses together with
a salt string, and iterating a given amount of times.

By default, the salt is random and not given to the user, and the number
of iterations is calculated by hashing until 100ms pass. This means that
the hashes are different on every run, and the IP address ban list is only
valid so long the server keeps running.